task: moved date config to a separate file & added daterange modal
